Resolve "Create an API call, with which to post questions"

Closes #38

See merge request cse1105/2020-2021/team-repositories/oopp-group-54/repository-template!28

+++ server/src/main/java/nl/tudelft/oopp/group54/controllers/questions/MockQuestionServiceImplementation.java
@@ -19,10 +19,10 @@ public class MockQuestionServiceImplementation implements QuestionService {
    * @param lectureId
    * @param userId
    * @param questionText
-   * @return status of the requst
+   * @return status of the request
    */
   @Override
-  public Map<String, Object> postQuestion(Long lectureId, String userId, String questionText) {
+  public Map<String, Object> postQuestion(Integer lectureId, String userId, String questionText) {
 
     Map<String, Object> status = new TreeMap<>();
     QuestionKey mockQuestionKey = new QuestionKey(123, 123);
